+++ src/main/java/itlu/itlu/controller/WorkerController.java
@@ -40,6 +40,9 @@ public class WorkerController {
 
     @GetMapping(path = "/{id}/deleteWorker")
     public String deleteDoctor(@PathVariable Long id) {
+        if(workerService.checkWorkerHasTeam(id)){
+            return "redirect:/canNotDeleteTeam";
+        }
         workerService.deleteWorker(id);
         return "redirect:/allEmployees";
     }
